+++ Runtime/Unity.Properties/PropertyPath.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Text;
 
@@ -20,8 +21,14 @@ namespace Unity.Properties
         public Part this[int index] => m_Parts[index];
 
         public PropertyPath()
+            :this(string.Empty)
+        {
+        }
+
+        public PropertyPath(string path)
         {
             m_Parts = new List<Part>(32);
+            ConstructFromPath(path);   
         }
 
         public void Push(string name, int index = InvalidListIndex)
@@ -55,19 +62,66 @@ namespace Unity.Properties
                 return string.Empty;
             }
 
-            var builder = new StringBuilder(16);
+            var builder = new StringBuilder(32);
 
             foreach (var part in m_Parts)
             {
-                if (!part.IsListItem && builder.Length > 0)
+                if (builder.Length > 0)
                 {
                     builder.Append('.');
                 }
 
                 builder.Append(part.Name);
+                if (part.IsListItem)
+                {
+                    builder.Append($"[{part.Index}]");
+                }
             }
 
             return builder.ToString();
+        }
+
+        private void ConstructFromPath(string path)
+        {
+            if (string.IsNullOrEmpty(path))
+            {
+                return;
+            }
+            
+            var parts = path.Split('.');
+            for (var i = 0; i < parts.Length; ++i)
+            {
+                var part = parts[i];
+
+                if (part.EndsWith("]"))
+                {
+                    var lastIndex = part.LastIndexOf("[", StringComparison.InvariantCultureIgnoreCase);
+                    if (lastIndex > 0)
+                    {
+                        var indexStr = part.Substring(lastIndex + 1, part.Length - 2 - lastIndex);
+                        if (int.TryParse(indexStr, out var index))
+                        {
+                            if (index < 0)
+                            {
+                                throw new ArgumentException($"Negative indices in {nameof(PropertyPath)} are not supported.");    
+                            }
+                            Push(part.Remove(lastIndex), index);
+                        }
+                        else
+                        {
+                            throw new ArgumentException($"Indices in {nameof(PropertyPath)} must be a numeric value.");
+                        }
+                    }
+                    else
+                    {
+                        throw new ArgumentException($"Indices may not be at the root of a {nameof(PropertyPath)}");
+                    }
+                }
+                else
+                {
+                    Push(part);
+                }
+            }
         }
     }
 }