+++ Runtime/Unity.Properties/PropertyVisitorAdapter.cs
@@ -265,7 +265,7 @@ namespace Unity.Properties
                 var elementChangeTracker = new ChangeTracker(changeTracker.VersionStorage);
                 var callback = new VisitCollectionElementCallback<TContainer>(visitor);
 
-                property.GetPropertyAtIndex(ref container, i, ref elementChangeTracker, callback);
+                property.GetPropertyAtIndex(ref container, i, ref elementChangeTracker, ref callback);
 
                 if (elementChangeTracker.IsChanged())
                 {