+++ Runtime/Unity.Properties/Reflection/Properties/ReflectedGenericListProperty.cs
@@ -78,7 +78,7 @@ namespace Unity.Properties.Reflection
 
         public void Clear(ref TContainer container) => GetValue(ref container).Clear();
 
-        public void GetPropertyAtIndex<TGetter>(ref TContainer container, int index, ref ChangeTracker changeTracker, TGetter getter) 
+        public void GetPropertyAtIndex<TGetter>(ref TContainer container, int index, ref ChangeTracker changeTracker, ref TGetter getter) 
             where TGetter : ICollectionElementPropertyGetter<TContainer>
         {
             getter.VisitProperty<ReflectedListElementProperty, TElement>(new ReflectedListElementProperty(this, index), ref container, ref changeTracker);